fix color picker

+++ src/libs/lib.h
@@ -47,9 +47,9 @@ typedef struct dt_lib_t
     struct
     {
       struct dt_lib_module_t *module;
-      uint8_t *picked_color_rgb_mean;
-      uint8_t *picked_color_rgb_min;
-      uint8_t *picked_color_rgb_max;
+      float *picked_color_rgb_mean;
+      float *picked_color_rgb_min;
+      float *picked_color_rgb_max;
       float *picked_color_lab_mean;
       float *picked_color_lab_min;
       float *picked_color_lab_max;
